ubusd: use fixed-width types for sequence counters
The message sequence number is a uint16_t on the wire (ubus_msghdr),
and ubusd_acl_seq is emitted as a u32. Declaring the counters with
the matching fixed-width types removes the implicit conversion that
happened on every store or send:

  event_seq               int          -> uint16_t
  ubus_object.invoke_seq  unsigned int -> uint16_t
  ubus_monitor.seq        uint32_t     -> uint16_t
  ubusd_acl_seq           int          -> uint32_t

The values placed on the wire are unchanged.

Link: https://github.com/openwrt/ubus/pull/20

+++ ubusd_event.c
@@ -16,7 +16,7 @@
 
 static struct avl_tree patterns;
 static struct ubus_object *event_obj;
-static int event_seq = 0;
+static uint16_t event_seq = 0;
 static int obj_event_seq = 1;
 
 struct event_source {
